test(breadcrumbs): migrate reactive
